requirements.txt
pywin32
Pillow
python-telegram-bot
pyautogui
pynput
colorlog
retrying
watchdog
PyGithub
tailer
